# docs/swagger.yaml
definitions:
  handlers.AddPredefinedTemplateRequest:
    properties:
      description:
        type: string
      event_type:
        type: string
      template:
        type: string
    type: object
  handlers.updateData:
    properties:
      channel_id:
        type: string
      description:
        type: string
    type: object
  handlers.updateWebhook:
    properties:
      webhook_origin:
        type: string
      webhook_secret:
        description: WebhookURL    string `json:"webhook_url"`
        type: string
    type: object
  models.DiscordMessage:
    properties:
      template:
        type: string
    type: object
  models.PredefinedTemplate:
    properties:
      created_at:
        type: string
      description:
        description: Optional description
        type: string
      event_type:
        description: Event type (e.g., "push", "pull_request")
        type: string
      id:
        type: integer
      template:
        description: Template content
        type: string
      updated_at:
        type: string
    type: object
  models.Project:
    properties:
      added_by:
        type: string
      channel_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      project_name:
        type: string
      project_source_auth_token:
        type: string
      project_source_url:
        type: string
      updated_at:
        type: string
      webhook_origin:
        type: string
      webhook_secret:
        type: string
      webhook_url:
        type: string
    type: object
  models.Template:
    properties:
      created_at:
        type: string
      description:
        description: Optional description
        type: string
      event_type:
        description: Event type (e.g., "push", "pull_request")
        type: string
      id:
        type: integer
      project_id:
        description: Foreign key to Project
        type: string
      template:
        description: Template content
        type: string
      updated_at:
        type: string
    type: object
  utils.Response:
    properties:
      data:
        type: object
      error:
        items:
          type: string
        type: array
      status:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /api/v1/predefined-templates:
    get:
      description: Retrieve all predefined templates
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.PredefinedTemplate'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Get all predefined templates
      tags:
      - PredefinedTemplates
    post:
      description: Create a new predefined template
      parameters:
      - description: Predefined template data
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/handlers.AddPredefinedTemplateRequest'
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PredefinedTemplate'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Add a new predefined template
      tags:
      - PredefinedTemplates
  /api/v1/predefined-templates/{templateID}:
    delete:
      description: Delete a predefined template by its ID
      parameters:
      - description: Template ID
        in: path
        name: templateID
        required: true
        type: string
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Delete a predefined template
      tags:
      - PredefinedTemplates
    get:
      description: Retrieve a predefined template by its ID
      parameters:
      - description: Template ID
        in: path
        name: templateID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PredefinedTemplate'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Get a predefined template by ID
      tags:
      - PredefinedTemplates
    put:
      description: Update the details of an existing predefined template
      parameters:
      - description: Template ID
        in: path
        name: templateID
        required: true
        type: string
      - description: Updated predefined template data
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/models.PredefinedTemplate'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PredefinedTemplate'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Update an existing predefined template
      tags:
      - PredefinedTemplates
  /api/v1/projects:
    get:
      consumes:
      - application/json
      description: Retrieves an array of all existing projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Project'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: List all projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a new project from the request body
      parameters:
      - description: Project to create
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Create a new project
      tags:
      - projects
  /api/v1/projects/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a project using the provided ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Delete a project by ID
      tags:
      - projects
    get:
      consumes:
      - application/json
      description: Retrieves a single project using the provided ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Get a project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Updates a project's details using the provided ID and request body
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated project data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/handlers.updateData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Update a project by ID
      tags:
      - projects
  /api/v1/projects/{id}/send-message:
    post:
      consumes:
      - application/json
      description: Sends a message to the Discord channel associated with the specified
        project ID. The message is rendered using a Go template provided in the request
        body.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Message payload with template
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.DiscordMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Send a message to a Discord channel
      tags:
      - projects
  /api/v1/projects/{id}/source:
    put:
      consumes:
      - application/json
      description: Updates a project's source authentication token using the provided
        ID and request body
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated project data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Update a project's source authentication token by ID
      tags:
      - projects
  /api/v1/projects/{id}/webhook:
    post:
      consumes:
      - application/json
      description: Generates a new webhook for a project using the provided ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Generate a new webhook for a project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Updates a project's webhook details using the provided ID and request
        body
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated project data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/handlers.updateWebhook'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Update a project's webhook details by ID
      tags:
      - projects
  /api/v1/projects/{id}/webhook/{webhookUrl}:
    post:
      consumes:
      - application/json
      description: Handles the webhook payload for a project using the provided ID
        and webhook URL
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Webhook URL
        in: path
        name: webhookUrl
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Handle webhook payload
      tags:
      - projects
  /api/v1/projects/{projectID}/templates:
    get:
      description: Retrieve all templates associated with a specific project
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Template'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Get all templates for a project
      tags:
      - Templates
    post:
      description: Create a new template for a project
      parameters:
      - description: Template data
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/models.Template'
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Template'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Add a new template
      tags:
      - Templates
  /api/v1/projects/{projectID}/templates/event-type/{eventType}:
    get:
      description: Retrieve a template associated with a specific project and event
        type
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Event type
        in: path
        name: eventType
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Template'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Get a template by event type
      tags:
      - Templates
  /api/v1/templates/{templateID}:
    delete:
      description: Delete a template by its ID
      parameters:
      - description: Template ID
        in: path
        name: templateID
        required: true
        type: string
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Delete a template
      tags:
      - Templates
    put:
      description: Update the details of an existing template
      parameters:
      - description: Template ID
        in: path
        name: templateID
        required: true
        type: string
      - description: Updated template data
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/models.Template'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Template'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Update an existing template
      tags:
      - Templates
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes the session token if the current session is valid.
      produces:
      - application/json
      responses:
        "200":
          description: Session refreshed successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Refresh the session token
      tags:
      - auth
swagger: "2.0"